move popper from dialogbody to messagebubble to simplify it
